requirements.txt
ccxt>=4.1.93
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiodns>=3.1.1
uvloop>=0.19.0; sys_platform != 'win32'